Documenting & thinking about a race condition, ZFLAG_WRITEOUT for txnmgr_atom_block_flush, working on copy-on-steal.
I accept terms and conditions stated in the Legal Agreement
(available at http://www.namesys.com/legalese.html)

2001/09/22 19:47:18-00:00 !jmacd
Some of the code in SLPC deals with allowing you to test in-kernel
code & data structures at user level.  Previously this was only
spinlocks, but an abstraction similar to kmem_slab_cache was needed
by the wandering log buffer manager, so I took this opportunity
to improve the separation of SLPC and its kernel-user test devices.
The libkut directory deals with these two abstractions.

I accept terms and conditions stated in the Legal Agreement
(available at http://www.namesys.com/legalese.html)

2001/09/15 22:16:11-00:00 !jmacd
I have finished running the testslpc2 test program on bowl which
is a 2-way SMP.  This allows me to compare the 1- and 2-process
cases which are now reported in the file RESULTS.

Overall the numbers help tune the data structure and the best node
size appears to be 256 bytes both with and without locking, with
and without concurrent access.  There are exceptions when the
size of the tree is very small (i.e., making everything fit in
one node), but that also means a high space overhead.

I am pleased to report that the effect of running 2 processes on
the same workload (i.e., the number of operations is held constant)
actually results in a speedup, meaning the concurrency does NOT hurt
performance.  Very cool.

The speedup is not dramatic, however.  Fixing the node at 256 bytes:

For a tree of approx 40000 elements the 2-processor case is 21% faster
For a tree of approx 400 elements the 2-processor case is 9% faster
For a tree of approx 40 elements the 2-processor case is 3% faster
For a tree of approx 4 ele

(Logical change 1.3)