Some measurements.
I accept terms and conditions stated in the Legal Agreement
(available at http://www.namesys.com/legalese.html)

2001/10/24 09:39:04-00:00 !jmacd
Documenting & thinking about a race condition, ZFLAG_WRITEOUT for
txnmgr_atom_block_flush, working on copy-on-steal.

I accept terms and conditions stated in the Legal Agreement
(available at http://www.namesys.com/legalese.html)

2001/10/01 16:42:01-00:00 !jmacd
A bug fix.

I accept terms and conditions stated in the Legal Agreement
(available at http://www.namesys.com/legalese.html)

2001/09/29 21:32:34-00:00 !jmacd
Playing around with GCC3.0.1.

I accept terms and conditions stated in the Legal Agreement
(available at http://www.namesys.com/legalese.html)

2001/09/23 20:35:21-00:00 !jmacd
Make a release 20010923.

I accept terms and conditions stated in the Legal Agreement
(available at http://www.namesys.com/legalese.html)

2001/09/22 19:47:18-00:00 !jmacd
Some of the code in SLPC deals with allowing you to test in-kernel
code & data structures at user level.  Previously this was only
spinlocks, but an abstraction similar to kmem_slab_cache was needed
by the wandering log buffer manager, so I took this opportunity
to improve the separation of SLPC and its kernel-user test devices.
The libkut directory deals with these two abstractions.

I accept terms and conditions stated in the Legal Agreement
(available at http://www.namesys.com/legalese.html)

2001/09/15 22:16:11-00:00 !jmacd
I have finished running the testslpc2 test program on bowl which
is a 2-way SMP.  This allows me to compare the 1- and 2-process
cases which are now reported in the file RESULTS.

Overall the numbers help tune the data structure and the best node
size appears to be 256 bytes both with and without locking, with
and without concurrent access.  There are exceptions when the
size of the tree is very small (i.e., making everything fit in
one node), but that also means a high space overhead.

I am pleased to report that the effect of running 2 processes on
the same workload (i.e., the number of operations is held constant)
actually results in a speedup, meaning the concurrency does NOT hurt
performance.  Very cool.

The speedup is not dramatic, however.  Fixing the node at 256 bytes:

For a tree of approx 40000 elements the 2-processor case is 21% faster
For a tree of approx 400 elements the 2-processor case is 9% faster
For a tree of approx 40 elements the 2-processor case is 3% faster
For a tree of approx 4 ele

2001/09/15 15:01:07-00:00 !jmacd
The multi-processor test now works.  I found only one bug, which
was a CPP #define typo which caused the _cycle count not to be
incremented, which caused upgrade races to occur.

I accept terms and conditions stated in the Legal Agreement
(available at http://www.namesys.com/legalese.html)

2001/09/15 00:18:06-00:00 !jmacd
More work on the multi-process SLPC test.  The 2-process test
appears to deadlock in slpc_upgrade_3locks, except that it
doesn't look like deadlock, it just looks like the read-write
spinlock isn't working.  More testing needed.

I accept terms and conditions stated in the Legal Agreement
(available at http://www.namesys.com/legalese.html)

2001/09/14 23:22:56-00:00 !jmacd
Add a new multi-processor test of SLPC.

I accept terms and conditions stated in the Legal Agreement
(available at http://www.namesys.com/legalese.html)

2001/09/13 13:58:52-00:00 !jmacd
This is all of the SLPC (skip list) code that I sent to the group
one month ago.  Included is a basic timing study that shows: (1) the
effect of matching the node size to the cache line size and (2) the
overhead of the spinlock functions as a function of node size.  The
results of the 'testslpc' application have been collected into the
file RESULTS.  Some of the code in this directory is taken from the
standard Linux 2.4 spinlocks, which require slight modification to
compile & test in user-space.  Those files retain their respective
copyright, and will not be necessary once this code was linked into
the kernel.

I accept terms and conditions stated in the Legal Agreement
(available at http://www.namesys.com/legalese.html)

(Logical change 1.3)