Some of the code in SLPC deals with allowing you to test in-kernel code & data structures at user level.  Previously this was only spinlocks, but an abstraction similar to kmem_slab_cache was needed by the wandering log buffer manager, so I took this opportunity to improve the separation of SLPC and its kernel-user test devices. The libkut directory deals with these two abstractions.
I accept terms and conditions stated in the Legal Agreement
(available at http://www.namesys.com/legalese.html)

2001/09/15 15:01:07-00:00 !jmacd
The multi-processor test now works.  I found only one bug, which
was a CPP #define typo which caused the _cycle count not to be
incremented, which caused upgrade races to occur.

I accept terms and conditions stated in the Legal Agreement
(available at http://www.namesys.com/legalese.html)

2001/09/13 13:58:52-00:00 !jmacd
This is all of the SLPC (skip list) code that I sent to the group
one month ago.  Included is a basic timing study that shows: (1) the
effect of matching the node size to the cache line size and (2) the
overhead of the spinlock functions as a function of node size.  The
results of the 'testslpc' application have been collected into the
file RESULTS.  Some of the code in this directory is taken from the
standard Linux 2.4 spinlocks, which require slight modification to
compile & test in user-space.  Those files retain their respective
copyright, and will not be necessary once this code was linked into
the kernel.

I accept terms and conditions stated in the Legal Agreement
(available at http://www.namesys.com/legalese.html)

(Logical change 1.3)