reiser4[format41]: jnode loading with failover - handle races
jload() became more complicated because of the feature of failover.
Now in the case of IO errors, or if jnode parsing failed, we restart
IO against replica devices, so we can not handle races by the old
way (without the feature of failover), when two processes loading
the same jnode issue IO and parse the results independently.

Now before jnode loading we first checks, if someone has already
started to load the jnode. If yes, then we wait for the completion
and return results of _that_ loading. Otherwise, we load and parse
the jnode by ourselves.

The sleeping is non-interruptable without timeouts. For simplicity
the wait queue is per-jnode. Later we can implement more economical
solutions where the queue is per-subvolume, or allocate per-jnode
queue by demand, so that jnode will contain only a pointer to the
queue (not the whole structure wait_queue_head, as it takes place
currently).

Signed-off-by: Edward Shishkin <edward.shishkin@gmail.com>