Fix up managing volinfo nodes: . use flag BA_PERMANENT to allocate blocks for volinfo nodes; . mark newly allocated volinfo nodes as "relocate"; . create an atom before allocating blocks for volmap nodes; . fixup up processing volinfo nodes at flush time.
Add a new bit field .init_vol_failed to reiser4_context to
indicate that volume initialization is not completed because
of errors, so that we don't need to commit transactions at
->release_format().

Handle cases of incomplete set of bricks registered in the
system when trying to mount a logical volume.

Starting to get rid of hardcoded assumption in the unix-file
plugin that the function (key -> offset-in-file-body) is
necessarily monotonic. Fixed can_append(), append_key_extent(),
insert_first_extent() and some assertions. Don't use file hints
to avoid tree traversal in the case when volume is not simple
(this is a stopgap measure for simplicity - later hints should
be adjusted also for compound volumes).

Signed-off-by: Edward Shishkin <edward.shishkin@gmail.com>