Fix up handle_pos_on_twig(): at the beginning of each iteration check data subvolume pointed out by extent item and make sure that we don't process more than one data subvolume per flush session.
Fix up the flush scan procedure: always look for the parent
in meta-data subvolume. Otherwise we'll fail to infinite
loop. Once we support more than one meta-data subvolumes,
just assign the proper one to look for the parent in.

Fix up find_or_create_extent():
call check_insert_atom_brick_info() after capturing a node
(not before) to make sure that atom already exists.

Fix up can_append() and append_hole() to get rid of
assumption that funcion (key -> ofset-in-file-body) is
necessarily monotonic. Update file size timely so that
can_append() is able to evaluate holes properly.

Fix up append_last_extent() to insert a new item (not simply
a unit) every time when a new stripe is going to be inserted.

With the above fixups we are able (by using sequential write)
to create a file with data scattered on different bricks.
Removing a file doesn't work, as cut_file_items() relies on
monotone of the function (key -> off-in-file-body) - to be
fixed later.

Signed-off-by: Edward Shishkin <edward.shishkin@gmail.com>