Roll back the previous commit because of design problems.
Representing holes as "fake" block pointers ("classic" design)
leads to scalability problems. In such design we can not merge
those pointers at stripe boundaries (otherwise we won't be able
to find them by a regular tree search procedure). Thus, when
creating a large hole (~1T) we'll need to insert a lot of block
pointers (~1000), which is unacceptable (currently creating a
large hole is a very fast op).

The single solution is to not represent holes by any items
(this approach is implemented in cryptcompress file plugin,
which doesn't manage the extent pointers though).

Signed-off-by: Edward Shishkin <edward.shishkin@gmail.com>