Keep a track of read-only state in reiser4 context
Added a new bit field .ro in reiser4 context to indicate
that execution started in read-only file system, so that
disk space wasn't grabbed, and we don't need to update
file's atime in that context.

This prevents an oops at boot time for a system with root
over reiser4 partition reported by Pierre-Philipp Braun.

Signed-off-by: Edward Shishkin <edward.shishkin@gmail.com>