Capture all super-blocks after adding/removing a brick
The alternative solution is storing all config parameters
(e.g. total number of bricks in the volume) in meta-data
brick and making sure that meta-data brick is activated
before all data bricks (it is more complicated, but will
provide better scalability).

Ensure correct order of writes of unbalanced and balanced
status of the volume.

Signed-off-by: Edward Shishkin <edward.shishkin@gmail.com>