Handle races with concurrent volume operations.
Under a long-term lock check if disk space reservation, made
on some data brick, is still valid, i.e. not spoiled by a "publisher"
(a concurrent operation on a logical volume ) after publishing a
new volume configuration, so that it turns out that "subscriber"
(a regular file operation) had to reserve space on another brick.

If reservation became invalid, then release the lock and make
reservation on a proper data brick.

Signed-off-by: Edward Shishkin <edward.shishkin@gmail.com>