Improved builtin extent migration procedure
Keep a track of migration progress by maintaining @done_off -
offset of the last successfully migrated byte. The trouble is
that we need to release the longterm lock for write throttling
per migrated page, so the item we are migrating can be changed.
We handle those cases by maintaining @stop_off (left stripe
boundary, which is invariant with respect to the locks).

Signed-off-by: Edward Shishkin <edward.shishkin@gmail.com>