Optimize data migration and operations on "striped" extents
Implement a new tree operation update_item_key(), which calls
carry to update delimiting keys on the upper tree levels. Earlier it
wasn't needed, because COP_UPDATE was always posted in the context of
carry operations (e.h. carry_cut()). Now specifics of work with
"striped" extents requires to update item keys by high level tree
operations.

Don't use kill-and-create-item sequence when migrating a whole extent,
as it requires to unlock current position in the tree, when converting
the first item in the file (suboptimal). Instead, convert the extent
to one unallocated unit, and update its key by the newly introduced
update_item_key()

In plug_hole_stripe() try also to push the pointer to the beginning
of the item at the right. If possible, update key of the item calling
update_item_key().

Signed-off-by: Edward Shishkin <edward.shishkin@gmail.com>